+++ src/java/lang/Long.h
@@ -1,14 +1,12 @@
 #ifndef __JAVA_LONG__
 #define __JAVA_LONG__
 
-#include <cstdint>
-
 namespace java {
 
 class Long {
   public:
-    static const int64_t MIN_VALUE = (-9223372036854775807LL - 1LL);
-    static const int64_t MAX_VALUE = 9223372036854775807LL;
+    static const long long MIN_VALUE = (-9223372036854775807LL - 1LL);
+    static const long long MAX_VALUE = 9223372036854775807LL;
 };
 
 }
